Ask about running with admin rights also when "run with powershell" #389
Windows10Debloater really needs to run with administrative rights,
otherwise it cannot ultimately remove the apps.

The problem was that the prompt to run as administrator
only worked within the ISE. It did not work if you
right-click on the ps1 and "run with PowerShell"
on a new installation of Windows.

This second case is by far the only way I would run a ps1,
and so it was always silently starting as a regular user,
and thus was largely ineffecive in actually removing bloat.

The host.ui displays a GUI prompt if available,
or a text prompt if not.
It also nicely allows setting a default value,
which I set to run as administrator.